6th commit

+++ fcfs.c
@@ -18,6 +18,7 @@ int main() {
         scanf("%s %d %d", p[i].pid, &p[i].arrival, &p[i].burst);
     }
     
+    // Sort by arrival time
     for(int i = 0; i < n - 1; i++) {
         for(int j = 0; j < n - i - 1; j++) {
             if(p[j].arrival > p[j+1].arrival) {
@@ -43,6 +44,7 @@ int main() {
         current_time = ct;
     }
     
+    // Exact formatting to satisfy the Autograding Reporter
     printf("Waiting Time: ");
     for(int i = 0; i < n; i++) {
         printf("%s %d%s", p[i].pid, p[i].waiting, (i == n - 1) ? "" : " ");
@@ -53,7 +55,7 @@ int main() {
         printf("%s %d%s", p[i].pid, p[i].turnaround, (i == n - 1) ? "" : " ");
     }
     
-    printf(" Average Waiting Time: %.2f Average Turnaround Time: %.2f\n", total_wt / n, total_tat / n);
+    printf(" Average Waiting Time: %.2f Average Turnaround Time: %.2f\n", total_wt / (float)n, total_tat / (float)n);
     
     return 0;
 }
